refactor(image form): migrate to typescript, this is a pain
