+++ packages/slice-machine/components/SliceMachineIconButton/index.tsx
@@ -20,8 +20,8 @@ const defaultActiveSx = (active: boolean, hasError: boolean | null): object => {
       };
 };
 
-type SliceMachineIconButtonProps = IconButtonProps & {
-  label: string;
+export type SliceMachineIconButtonProps = IconButtonProps & {
+  label?: string;
   size?: number;
   error?: boolean;
   fitButton?: boolean;
